make useTriggerElement generic
